+++ selfdrive/car/chrysler/carstate.py
@@ -4,11 +4,14 @@ from opendbc.can.can_define import CANDefine
 from selfdrive.config import Conversions as CV
 from selfdrive.car.interfaces import CarStateBase
 from selfdrive.car.chrysler.values import DBC, STEER_THRESHOLD
-from common.op_params import opParams
+from common.cached_params import CachedParams
 
 ButtonType = car.CarState.ButtonEvent.Type
 
-LEAD_RADAR_CONFIG = ['lead_distance_ratio_1bar', 'lead_distance_ratio_2bars', 'lead_distance_ratio_3bars', 'lead_distance_ratio_4bars']
+LEAD_RADAR_CONFIG = ['jvePilot.settings.accFollow1RadarRatio',
+                     'jvePilot.settings.accFollow2RadarRatio',
+                     'jvePilot.settings.accFollow3RadarRatio',
+                     'jvePilot.settings.accFollow4RadarRatio']
 CHECK_BUTTONS = {ButtonType.cancel: 'ACC_CANCEL',
                  ButtonType.resumeCruise: 'ACC_RESUME',
                  ButtonType.accelCruise: 'ACC_SPEED_INC',
@@ -19,12 +22,12 @@ CHECK_BUTTONS = {ButtonType.cancel: 'ACC_CANCEL',
 class CarState(CarStateBase):
   def __init__(self, CP):
     super().__init__(CP)
-    self.op_params = opParams()
+    self.cachedParams = CachedParams()
     can_define = CANDefine(DBC[CP.carFingerprint]['pt'])
     self.shifter_values = can_define.dv["GEAR"]['PRNDL']
 
   def update(self, cp, cp_cam):
-    speed_adjust_ratio = self.op_params.get('op_speed_adjust_ratio')
+    speed_adjust_ratio = self.cachedParams.get_float('jvePilot.settings.speedAdjustRatio', 5000)
     inverse_speed_adjust_ratio = 2 - speed_adjust_ratio
 
     ret = car.CarState.new_message()
@@ -76,7 +79,7 @@ class CarState(CarStateBase):
     ret.genericToggle = bool(cp.vl["STEERING_LEVERS"]['HIGH_BEAM_FLASH'])
 
     ret.jvePilotCarState.accFollowDistance = int(min(3, max(0, cp.vl["DASHBOARD"]['ACC_DISTANCE_CONFIG_2'])))
-    ret.jvePilotCarState.leadDistanceRadarRatio = self.op_params.get(LEAD_RADAR_CONFIG[ret.jvePilotCarState.accFollowDistance]) * inverse_speed_adjust_ratio
+    ret.jvePilotCarState.leadDistanceRadarRatio = self.cachedParams.get_float(LEAD_RADAR_CONFIG[ret.jvePilotCarState.accFollowDistance], 1000) * inverse_speed_adjust_ratio
 
     self.lkas_counter = cp_cam.vl["LKAS_COMMAND"]['COUNTER']
     self.lkas_car_model = cp_cam.vl["LKAS_HUD"]['CAR_MODEL']