+++ selfdrive/controls/lib/lane_planner.py
@@ -2,19 +2,18 @@ from common.numpy_fast import interp
 import numpy as np
 from selfdrive.hardware import EON, TICI
 from cereal import log
-from common.op_params import opParams
-
+from common.cached_params import CachedParams
 
 TRAJECTORY_SIZE = 33
 # camera offset is meters from center car to camera
 if EON:
-  #CAMERA_OFFSET = 0.06
+  CAMERA_OFFSET = 0.06
   PATH_OFFSET = 0.0
 elif TICI:
-  #CAMERA_OFFSET = -0.04
+  CAMERA_OFFSET = -0.04
   PATH_OFFSET = -0.04
 else:
-  #CAMERA_OFFSET = 0.0
+  CAMERA_OFFSET = 0.0
   PATH_OFFSET = 0.0
 
 
@@ -39,7 +38,7 @@ class LanePlanner:
     self.l_lane_change_prob = 0.
     self.r_lane_change_prob = 0.
 
-    self.op_params = opParams()
+    self.cachedParams = CachedParams()
 
 
   def parse_model(self, md):
@@ -48,9 +47,9 @@ class LanePlanner:
       # left and right ll x is the same
       self.ll_x = md.laneLines[1].x
       # only offset left and right lane lines; offsetting path does not make sense
-      CAMERA_OFFSET = self.op_params.get('camera_offset')
-      self.lll_y = np.array(md.laneLines[1].y) - CAMERA_OFFSET
-      self.rll_y = np.array(md.laneLines[2].y) - CAMERA_OFFSET
+      device_offset = self.cachedParams.get_float('jvePilot.settings.deviceOffset', 5000)
+      self.lll_y = np.array(md.laneLines[1].y) - CAMERA_OFFSET + device_offset
+      self.rll_y = np.array(md.laneLines[2].y) - CAMERA_OFFSET + device_offset
       self.lll_prob = md.laneLineProbs[1]
       self.rll_prob = md.laneLineProbs[2]
       self.lll_std = md.laneLineStds[1]