+++ selfdrive/ui/qt/widgets/controls.hpp
@@ -11,11 +11,22 @@
 
 QFrame *horizontal_line(QWidget *parent = nullptr);
 
+// config for a widget
+struct ConfigButton {
+  public:
+  std::string param;
+  std::string default_value;
+  float min;
+  float max;
+  QString title;
+  QString text; 
+};
+
 class AbstractControl : public QFrame {
   Q_OBJECT
 
 protected:
-  AbstractControl(const QString &title, const QString &desc = "", const QString &icon = "", QWidget *parent = nullptr);
+  AbstractControl(const QString &title, const QString &desc = "", const QString &icon = "", QList<struct ConfigButton> *btns = {}, QWidget *parent = nullptr);
 
   QSize minimumSizeHint() const override {
     QSize size = QFrame::minimumSizeHint();
@@ -33,7 +44,7 @@ class LabelControl : public AbstractControl {
   Q_OBJECT
 
 public:
-  LabelControl(const QString &title, const QString &text, const QString &desc = "", QWidget *parent = nullptr) : AbstractControl(title, desc, "", parent) {
+  LabelControl(const QString &title, const QString &text, const QString &desc = "", QList<struct ConfigButton> *btns = {}, const QString &icon = "", QWidget *parent = nullptr) : AbstractControl(title, desc, icon, btns, parent) {
     label.setText(text);
     label.setAlignment(Qt::AlignRight | Qt::AlignVCenter);
     hlayout->addWidget(&label);
@@ -50,7 +61,7 @@ class ButtonControl : public AbstractControl {
 
 public:
   template <typename Functor>
-  ButtonControl(const QString &title, const QString &text, const QString &desc, Functor functor, const QString &icon = "", QWidget *parent = nullptr) : AbstractControl(title, desc, icon, parent) {
+  ButtonControl(const QString &title, const QString &text, const QString &desc, Functor functor, const QString &icon = "", QList<struct ConfigButton> *btns = {}, QWidget *parent = nullptr) : AbstractControl(title, desc, icon, btns, parent) {
     btn.setText(text);
     btn.setStyleSheet(R"(
       QPushButton {
@@ -69,7 +80,10 @@ public:
     QObject::connect(&btn, &QPushButton::released, functor);
     hlayout->addWidget(&btn);
   }
+  void setLabel(const QString &text) { title_label->setText(text); }
   void setText(const QString &text) { btn.setText(text); }
+  template <typename Functor>
+  void released(Functor functor) { QObject::connect(&btn, &QPushButton::released, functor); }
 
 public slots:
   void setEnabled(bool enabled) {
@@ -84,7 +98,7 @@ class ToggleControl : public AbstractControl {
   Q_OBJECT
 
 public:
-  ToggleControl(const QString &title, const QString &desc = "", const QString &icon = "", const bool state = false, QWidget *parent = nullptr) : AbstractControl(title, desc, icon, parent) {
+  ToggleControl(const QString &title, const QString &desc = "", const QString &icon = "", const bool state = false, QList<struct ConfigButton> *btns = {}, QWidget *parent = nullptr) : AbstractControl(title, desc, icon, btns, parent) {
     toggle.setFixedSize(150, 100);
     if (state) {
       toggle.togglePosition();
@@ -107,7 +121,7 @@ class ParamControl : public ToggleControl {
   Q_OBJECT
 
 public:
-  ParamControl(const QString &param, const QString &title, const QString &desc, const QString &icon, QWidget *parent = nullptr) : ToggleControl(title, desc, icon, parent) {
+  ParamControl(const QString &param, const QString &title, const QString &desc, const QString &icon, QList<struct ConfigButton> *btns = {}, QWidget *parent = nullptr) : ToggleControl(title, desc, icon, false, btns, parent) {
     // set initial state from param
     if (Params().read_db_bool(param.toStdString().c_str())) {
       toggle.togglePosition();