+++ RELEASES.md
@@ -1,10 +1,14 @@
+Version jvePilot-0.8.3(v3.1) (2021-05-04)
+========================
+* Move everything to the UI. No more opEdit. All settings will need to be reset in the UI.
+* Curvature drop off setting 
+
 Version jvePilot-0.8.3(v3.0) (2021-04-22)
 ========================
 * ACC Eco
 * Fix Auto Follow ping/pong
 * Code changes to better consolidate state/control messaging
 * Option to reverse 1mph or 5mph on short ACC +/- presses 
-
 
 Version jvePilot-0.8.3(v2.2) (2021-04-11)
 ========================